Bump to v0.5.0
Incorporates Lunr 2.1.5 and includes Jekyll 3.x compatibility fixes.

Replaces inline JavaScript with a separate script to set constants
generated at site build time. This enables the use of the
Content-Security-Policy header to prevent the execution of inline
JavaScript (previously a `script-src 'unsafe-inline'` exception was
required to use the search feature).